+++ packages/remme-batch/src/interface.ts
@@ -1,5 +1,0 @@
-import { BatchResult } from "./models";
-
-export interface IRemmeBatch {
-    getStatus(batchId: string): Promise<BatchResult>;
-}