+++ packages/remme-blockchain-info/src/index.ts
@@ -17,6 +17,7 @@ import {
     BlockData,
     BlockList,
     State,
+    StateData,
     StateList,
     StateQuery,
     Transaction,
@@ -31,6 +32,7 @@ import {
     IBlockInfoResponse,
     IBlockInfo,
     BlockInfo,
+    IAddress,
 } from "./models";
 
 class RemmeBlockchainInfo implements IRemmeBlockchainInfo {
@@ -111,7 +113,7 @@ class RemmeBlockchainInfo implements IRemmeBlockchainInfo {
         }
     }
 
-    private _prepareAddress(state): State {
+    private _prepareAddress(state: State): State {
         if (RemmeBlockchainInfo.address[state.address.slice(0, 6)]) {
             const { parser: protobuf, type: addressType } = RemmeBlockchainInfo.address[state.address.slice(0, 6)];
             return {
@@ -159,7 +161,7 @@ class RemmeBlockchainInfo implements IRemmeBlockchainInfo {
 
     public async getBatchById(id: string): Promise<Batch> {
         this._checkId(id);
-        const apiResult = await this._remmeRest.getRequest<Batch>(RemmeMethods.fetchBatch, { id });
+        const apiResult = await this._remmeRest.sendRequest<IAddress, Batch>(RemmeMethods.fetchBatch, { id });
         apiResult.data = this._prepareBatch(apiResult.data);
         return apiResult;
     }
@@ -168,59 +170,60 @@ class RemmeBlockchainInfo implements IRemmeBlockchainInfo {
         if (query) {
             query = new BaseQuery(query);
         }
-        const apiResult = await this._remmeRest.getRequest<BatchList>(RemmeMethods.batches, query);
-        apiResult.data = apiResult.data.map((item) => {
-            return this._prepareBatch(item);
-        });
+        const apiResult = await this._remmeRest.sendRequest<BaseQuery, BatchList>(RemmeMethods.batches, query);
+        apiResult.data = apiResult.data.map((item) => this._prepareBatch(item));
         return apiResult;
     }
 
     public async getBlockById(id: string): Promise<Block> {
         this._checkId(id);
-        const apiResult = await this._remmeRest.getRequest<Block>(RemmeMethods.fetchBlock, { id });
+        const apiResult = await this._remmeRest.sendRequest<IAddress, Block>(RemmeMethods.fetchBlock, { id });
         apiResult.data = this._prepareBlock(apiResult.data);
         return apiResult;
     }
 
     public async getBlocks(query?: IBaseQuery): Promise<BlockList> {
         if (query) {
+            if (typeof query.start === "number") {
+                query.start = `0x${("0000000000000000" + query.start.toString(16)).slice(-16)}`;
+            }
             query = new BaseQuery(query);
         }
-        const apiResult = await this._remmeRest.getRequest<BlockList>(RemmeMethods.blocks, query);
-        apiResult.data = apiResult.data.map((block) => {
-            return this._prepareBlock(block);
-        });
+        const apiResult = await this._remmeRest.sendRequest<BaseQuery, BlockList>(RemmeMethods.blocks, query);
+        apiResult.data = apiResult.data.map((block) => this._prepareBlock(block));
         return apiResult;
     }
 
     public async getPeers(): Promise<PeerList> {
-        return await this._remmeRest.getRequest<PeerList>(RemmeMethods.peers);
+        return await this._remmeRest.sendRequest<PeerList>(RemmeMethods.peers);
     }
 
     public async getReceipts(ids: string[]): Promise<ReceiptList> {
         ids.map((id) => this._checkId(id));
-        return await this._remmeRest.getRequest<ReceiptList>(RemmeMethods.receipts, { ids });
+        return await this._remmeRest.sendRequest<IAddress, ReceiptList>(RemmeMethods.receipts, { ids });
     }
 
     public async getState(query?: IStateQuery): Promise<StateList> {
         if (query) {
             query = new StateQuery(query);
         }
-        const apiResult = await this._remmeRest.getRequest<StateList>(RemmeMethods.state, query);
+        const apiResult = await this._remmeRest.sendRequest<StateQuery, StateList>(RemmeMethods.state, query);
         apiResult.data = apiResult.data.map((state) => this._prepareAddress(state));
         return apiResult;
     }
 
     public async getStateByAddress(address: string): Promise<State> {
         this._checkAddress(address);
-        let apiResult = await this._remmeRest.getRequest<State>(RemmeMethods.fetchState, { address });
+        let apiResult = await this._remmeRest
+            .sendRequest<IAddress, State>(RemmeMethods.fetchState, { address });
         apiResult = this._prepareAddress({ address, ...apiResult });
         return apiResult;
     }
 
     public async getTransactionById(id: string): Promise<Transaction> {
         this._checkId(id);
-        const apiResult = await this._remmeRest.getRequest<Transaction>(RemmeMethods.fetchTransaction, { id });
+        const apiResult = await this._remmeRest
+            .sendRequest<IAddress, Transaction>(RemmeMethods.fetchTransaction, { id });
         apiResult.data = this._prepareTransaction(apiResult.data);
         return apiResult;
     }
@@ -229,7 +232,8 @@ class RemmeBlockchainInfo implements IRemmeBlockchainInfo {
         if (query) {
             query = new BaseQuery(query);
         }
-        const apiResult = await this._remmeRest.getRequest<TransactionList>(RemmeMethods.transactions, query);
+        const apiResult = await this._remmeRest
+            .sendRequest<BaseQuery, TransactionList>(RemmeMethods.transactions, query);
         apiResult.data = apiResult.data.map((item) => {
             return this._prepareTransaction(item);
         });
@@ -238,13 +242,13 @@ class RemmeBlockchainInfo implements IRemmeBlockchainInfo {
 
     public async getNetworkStatus(): Promise<INetworkStatus> {
         const apiResult = await this._remmeRest
-            .getRequest<INetworkStatusResponse>(RemmeMethods.networkStatus);
+            .sendRequest<INetworkStatusResponse>(RemmeMethods.networkStatus);
         return new NetworkStatus(apiResult);
     }
 
     public async getBlockInfo(query?: IBaseQuery): Promise<IBlockInfo[]> {
         const blocks = await this._remmeRest
-            .getRequest<IBlockInfoResponse[]>(RemmeMethods.blockInfo, query);
+            .sendRequest<IBaseQuery, IBlockInfoResponse[]>(RemmeMethods.blockInfo, query);
         if (!blocks) {
             throw new Error("Unknown error occurs in the server");
         }
@@ -253,7 +257,33 @@ class RemmeBlockchainInfo implements IRemmeBlockchainInfo {
 
     public async getBatchStatus(id: string): Promise<string> {
         this._checkId(id);
-        return await this._remmeRest.getRequest<string>(RemmeMethods.batchStatus, { id });
+        return await this._remmeRest.sendRequest<IAddress, string>(RemmeMethods.batchStatus, { id });
+    }
+
+    public parseTransactionPayload(transaction: Transaction): object {
+        const { family_name } = transaction.header;
+        if (family_name in RemmeBlockchainInfo.correspond) {
+            const { method, data } = protobufs.TransactionPayload.decode(base64ToArrayBuffer(transaction.payload));
+            const { parser, type } = RemmeBlockchainInfo.correspond[family_name][method];
+            return {
+                payload: parser.decode(data),
+                type,
+            };
+        } else {
+            throw new Error(`This family name (${family_name}) don't supported for parsing`);
+        }
+    }
+
+    public parseStateData(state: State): object {
+        if (RemmeBlockchainInfo.address[state.address.slice(0, 6)]) {
+            const { parser, type } = RemmeBlockchainInfo.address[state.address.slice(0, 6)];
+            return {
+                data: parser.decode(base64ToArrayBuffer(state.data)),
+                type,
+            };
+        } else {
+            throw new Error(`This address (${state.address}) don't supported for parsing`);
+        }
     }
 }