+++ packages/remme-public-key-storage/src/models/PublicKeyStorageCheckResult.ts
@@ -1,7 +1,7 @@
 export class PublicKeyStorageCheckResult {
-    public owner: string;
-    public revoked: boolean;
-    public valid: boolean;
+    public owner_public_key: string;
+    public is_revoked: boolean;
+    public is_valid: boolean;
     public valid_from: number;
     public valid_to: number;
     public entity_hash: string;