+++ packages/remme/src/interface.ts
@@ -1,7 +1,6 @@
 import { IRemmeCertificate } from "remme-certificate";
 import { IRemmeToken } from "remme-token";
 import { IRemmeAccount } from "remme-account";
-import { IRemmeBatch } from "remme-batch";
 import { IRemmeSwap } from "remme-atomic-swap";
 import { IRemmeTransactionService } from "remme-transaction-service";
 import { IRemmePublicKeyStorage } from "remme-public-key-storage";
@@ -13,7 +12,6 @@ export interface IRemmeClient {
     certificate: IRemmeCertificate;
     token: IRemmeToken;
     account: IRemmeAccount;
-    batch: IRemmeBatch;
     swap: IRemmeSwap;
     transaction: IRemmeTransactionService;
     publicKeyStorage: IRemmePublicKeyStorage;